Extract VolumeData interface
