+++ src/app/interfaces/events/volume-data-event.interface.ts
@@ -1,0 +1,8 @@
+import { ApiService } from 'app/core/services/api.service';
+import { VolumeData } from '../volume-data.interface';
+
+export interface VolumeDataEvent {
+  name: 'VolumeData';
+  sender: ApiService;
+  data: VolumeData[];
+}
