+++ src/app/interfaces/volume-data.interface.ts
@@ -1,0 +1,15 @@
+export interface VolumeData {
+  avail?: number;
+  id?: number;
+  is_decrypted?: boolean;
+  is_upgraded?: boolean;
+  mountpoint?: string;
+  name?: string;
+  status?: string;
+  used?: number;
+  used_pct?: string;
+  vol_encrypt?: number;
+  vol_encryptkey?: string;
+  vol_guid?: string;
+  vol_name?: string;
+}
